# docs/openapi.yaml
openapi: 3.0.3
info:
  title: dPagès API v1
  version: '1.0'
  description: >
    API interna del sistema de gestión de comandas de dPagès. Fuente única de
    verdad del contrato entre backend y frontend — ver también
    docs/contrato-api.md para las explicaciones de negocio detrás de cada
    regla (por qué el empaquetado exige valores obligatorios, qué es
    kgEditable, etc.).

    Refleja el estado confirmado hasta la capa 11 (detalle de incidencias,
    CORS, alta manual de cliente, dirección de entrega del pedido, fecha y
    observación por línea, código de tarifa) MÁS las decisiones funcionales
    confirmadas con el cliente el 18/08/2026: campos de agrupación de
    artículo (formato, envasado, agrupación de producción), agrupación de
    rendimiento por categoría, Rendiments Porcs, Panell Producció (shape
    cerrado; las fórmulas numéricas todavía no — ver
    docs/contrato-api.md sección 4.10), Panell Obrador por línea individual
    (ya no agrupado por producto), orígenes de pedido como tabla mantenible
    (ya no un enum fijo), y usuarios/roles. Los endpoints que en su momento
    eran sólo contrato (rendiments-porcs, panells/produccio, usuaris, rols,
    jo, origens-comanda) ya están implementados en el backend real, cada
    uno con su suite de tests.
  contact:
    name: VisioFlow Tech
servers:
  - url: http://localhost:8080
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/comandes:
    get:
      tags: [Comandes]
      summary: Listar pedidos
      parameters:
        - name: estat
          in: query
          schema: { $ref: '#/components/schemas/EstatComanda' }
        - name: clientId
          in: query
          schema: { type: integer }
        - name: origen
          in: query
          description: 'Codi de origen_comanda, ej. "woocommerce" o "manual" — ya no es un enum fijo.'
          schema: { type: string }
        - name: dataDes
          in: query
          description: Filtra por dataComanda (comanda.creat_en) — sin cambios.
          schema: { type: string, format: date }
        - name: dataFins
          in: query
          description: >-
            Filtra por dataComanda (comanda.creat_en) — sin cambios de
            criterio de comparación. Capa 36: incluye el día completo (hasta
            las 23:59:59.999 de esa fecha), no sólo su medianoche inicial.
          schema: { type: string, format: date }
        - name: dataProduccioDes
          in: query
          description: >-
            Capa 21. Matchea si AL MENOS UNA línea del pedido tiene
            dataProduccio dentro del rango (la misma línea cumple
            dataProduccioDes Y dataProduccioFins a la vez, no dos líneas
            distintas cumpliendo cada una un extremo).
          schema: { type: string, format: date }
        - name: dataProduccioFins
          in: query
          description: >-
            Ver dataProduccioDes. Capa 36: incluye el día completo, igual
            que el resto de los "...Fins" del proyecto.
          schema: { type: string, format: date }
        - name: dataLliuramentDes
          in: query
          description: Capa 21. Filtra por comanda.dataLliurament (cabecera).
          schema: { type: string, format: date }
        - name: dataLliuramentFins
          in: query
          description: >-
            Ver dataLliuramentDes. Capa 36: incluye el día completo, igual
            que el resto de los "...Fins" del proyecto.
          schema: { type: string, format: date }
        - name: cerca
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Mida'
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dades:
                    type: array
                    items: { $ref: '#/components/schemas/ComandaResum' }
                  paginacio: { $ref: '#/components/schemas/Paginacio' }
    post:
      tags: [Comandes]
      summary: Alta manual de pedido (teléfono, email, WhatsApp)
      description: >-
        Capa 34: linies[].dataProduccio (si viene) se valida contra las 6
        reglas de coherencia temporal (ver ComandaEntradaParcial.
        dataProduccio) — en este endpoint, como dataProduccio/dataExpedicio
        de cabecera no son campos de este body, en la práctica sólo puede
        dispararse la regla 5 (línea vs. dataLliurament).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComandaEntrada' }
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComandaDetall' }
        '400': { $ref: '#/components/responses/Validacio' }
  /api/v1/comandes/{id}:
    get:
      tags: [Comandes]
      summary: Detalle de un pedido, con líneas e incidencias
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Pedido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComandaDetall' }
        '404': { $ref: '#/components/responses/NoTrobat' }
    patch:
      tags: [Comandes]
      summary: Actualizar cabecera del pedido (parcial)
      description: >-
        Capa 31. Incluye `estat`: transición libre entre los 5 valores de
        EstatComanda, sin máquina de estados. Si `estat` es "amb_incidencia",
        `detall` es obligatorio en el mismo body (400 si falta) y se registra
        como una incidencia nueva (tipus "manual"). Capa 34: si cambia
        alguna fecha de cabecera, se valida contra las 6 reglas de
        coherencia temporal (ver ComandaEntradaParcial.dataProduccio),
        incluyendo TODAS las líneas activas del pedido, no sólo las que este
        request esté tocando (no toca ninguna). 409 si la comanda está
        congelada.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComandaEntradaParcial' }
      responses:
        '200':
          description: Pedido actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComandaDetall' }
        '400': { $ref: '#/components/responses/Validacio' }
        '404': { $ref: '#/components/responses/NoTrobat' }
        '409': { $ref: '#/components/responses/Conflicte' }
...
  /api/v1/usuaris:
    get:
      tags: [Usuaris]
      summary: Listar usuarios (capa 17).
      parameters:
        - name: actiu
          in: query
          schema: { type: boolean }
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Mida'
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  dades:
                    type: array
                    items: { $ref: '#/components/schemas/Usuari' }
                  paginacio: { $ref: '#/components/schemas/Paginacio' }
    post:
      tags: [Usuaris]
      summary: >-
        Alta manual de un usuario, hecha por un Administrador (capa 19).
      description: >-
        Restringido: sólo usuarios con "usuaris" en su modulsPermesos
        (403 SENSE_PERMIS si no). El backend no envía ningún email — crea el
        usuario en Firebase Authentication (con una contraseña aleatoria
        descartable), crea la fila local, genera con el Admin SDK de
        Firebase un link de "establecé tu contraseña" de un solo uso
        (generatePasswordResetLink) y lo devuelve en linkEstabliment. Quien
        dio de alta a la persona comparte ese link a mano (WhatsApp, email
        personal, etc.) — no hay ningún otro mecanismo de entrega. Si algo
        falla después de crear el usuario en Firebase, se revierte
        (deleteUser) para no dejar un usuario huérfano allá.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuariCreacio' }
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsuariCreatResposta' }
        '400': { $ref: '#/components/responses/Validacio' }
        '403': { $ref: '#/components/responses/SensePermis' }
        '409': { $ref: '#/components/responses/Conflicte' }
  /api/v1/usuaris/{id}:
    patch:
      tags: [Usuaris]
      summary: >-
        Actualizar un usuario (parcial) — sólo nom/rolId/actiu.
        firebaseUid y email son inmutables una vez creado.
      description: >-
        Capa 39: restringido, sólo usuarios con "usuaris" en su
        modulsPermesos (403 SENSE_PERMIS si no) — mismo guard que ya usaba
        POST /usuaris. Corrige un agujero de seguridad: cualquier usuario
        autenticado podía editar rolId de cualquiera, incluido el propio
        (auto-promoción a Administrador).
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UsuariEdicio' }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuari' }
        '400': { $ref: '#/components/responses/Validacio' }
        '403': { $ref: '#/components/responses/SensePermis' }
        '404': { $ref: '#/components/responses/NoTrobat' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: >
        Requerido en todas las rutas excepto /salut. En desarrollo local con
        AUTH_DISABLED=true no hace falta enviarlo.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    Pagina:
      name: pagina
      in: query
      schema: { type: integer, default: 1 }
    Mida:
      name: mida
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  responses:
    NoTrobat:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Validacio:
      description: Error de validación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflicte:
      description: Conflicto (p. ej. pedido congelado)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SensePermis:
      description: 'Autenticado pero sin permiso (ej. usuari amb actiu=false)'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    EstatComanda:
      type: string
      enum: [oberta, en_proces, tancada, amb_incidencia, cancellada]
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            codi:
              type: string
              enum: [VALIDACIO, NO_AUTENTICAT, SENSE_PERMIS, NO_TROBAT, CONFLICTE, ERROR_INTERN]
            missatge: { type: string }
            detalls:
              type: array
              items:
                type: object
                properties:
                  camp: { type: string }
                  missatge: { type: string }
    Paginacio:
      type: object
      properties:
        pagina: { type: integer }
        mida: { type: integer }
        total: { type: integer }
        totalPagines: { type: integer }
...
    ComandaResum:
      type: object
      description: >-
        Forma usada en GET /comandes. (Corrección capa 35: GET
        /panells/oficina usa FilaPanellOficina, una forma distinta — no
        ComandaResum, como decía antes esta descripción.)
      properties:
        id: { type: integer }
        num: { type: string }
        origen:
          type: string
          description: >-
            Codi de origen_comanda (ej. "woocommerce", "manual") — ya no es
            un enum fijo, ver el schema OrigenComanda y GET /origens-comanda.
        estat: { $ref: '#/components/schemas/EstatComanda' }
        client:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            nom: { type: string }
            poblacio: { type: string }
        tarifa:
          type: object
          nullable: true
          properties: { id: { type: integer }, nom: { type: string } }
        transportista:
          type: object
          nullable: true
          properties: { id: { type: integer }, nom: { type: string } }
        poblacioDesti: { type: string, nullable: true }
        dataComanda: { type: string, format: date-time }
        dataProduccio: { type: string, format: date-time, nullable: true }
        datesProduccioLinies:
          type: array
          items: { type: string, format: date-time }
          description: >-
            Capa 21. Fechas de producción DISTINTAS entre las líneas del
            pedido (ComandaLinia.dataProduccio), ordenadas
            cronológicamente, sin nulls. Array vacío (nunca null) si
            ninguna línea tiene fecha de producción. Distinto de
            dataProduccio (arriba, a nivel de cabecera).
        dataExpedicio: { type: string, format: date-time, nullable: true }
        dataLliurament: { type: string, format: date-time, nullable: true }
        bultos: { type: integer, nullable: true }
        totalLinies: { type: integer }
        totalKg: { type: string }
        totalEur: { type: string }
        congelada: { type: boolean }
        totalIncidencies:
          type: integer
          description: Cantidad de incidencias asociadas al pedido
        tipusIncidencia:
          type: string
          nullable: true
          description: Null si hay más de un tipo de incidencia mezclado
...
    Incidencia:
      type: object
      properties:
        tipus:
          type: string
          example: article_no_resolt
        detall: { type: string }
        creatA: { type: string, format: date-time }
    ComandaDetall:
      type: object
      description: Forma usada en GET /comandes/:id y POST /comandes
      properties:
        id: { type: integer }
        num: { type: string }
        origen:
          type: string
          description: >-
            Codi de origen_comanda (ej. "woocommerce", "manual") — ya no es
            un enum fijo, ver el schema OrigenComanda y GET /origens-comanda.
        estat: { $ref: '#/components/schemas/EstatComanda' }
        client:
          type: object
          nullable: true
          properties: { id: { type: integer }, nom: { type: string }, poblacio: { type: string } }
        tarifa:
          type: object
          nullable: true
          properties: { id: { type: integer }, nom: { type: string } }
        transportista:
          type: object
          nullable: true
          properties: { id: { type: integer }, nom: { type: string } }
        poblacioDesti: { type: string, nullable: true }
        adrecaLliurament:
          type: string
          nullable: true
          description: Dirección de entrega, distinta de la población de destino
        dataComanda: { type: string, format: date-time }
        dataProduccio: { type: string, format: date-time, nullable: true }
        dataExpedicio: { type: string, format: date-time, nullable: true }
        dataLliurament: { type: string, format: date-time, nullable: true }
        bultos: { type: integer, nullable: true }
        obsProduccio: { type: string, nullable: true }
        obsLliurament: { type: string, nullable: true }
        totalKg: { type: string }
        totalEur: { type: string }
        congelada: { type: boolean }
        congelatA: { type: string, format: date-time, nullable: true }
        linies:
          type: array
          items: { $ref: '#/components/schemas/ComandaLinia' }
        incidencies:
          type: array
          items: { $ref: '#/components/schemas/Incidencia' }
    ComandaEntrada:
      type: object
      required: [origen, clientId, linies]
      properties:
        origen:
          type: string
          description: >-
            Codi de origen_comanda (ej. "telefon", "whatsapp") — ya no es
            un enum fijo, ver el schema OrigenComanda y GET /origens-comanda.
            "woocommerce" no se puede elegir a mano (tarea 11): 400.
            Sólo lo asigna la sincronización.
        clientId: { type: integer }
        tarifaId:
          type: integer
          description: >-
            Capa 32. Opcional — si viene, anula la tarifa del cliente SÓLO
            para resolver el precio de las líneas de esta alta, y se guarda
            en comanda.tarifaId. Si no viene, se usa client.tarifaId (sin
            cambios respecto al comportamiento anterior a esta capa).
            Editarlo después vía PATCH /comandes/{id} no recalcula estas
            líneas, y POST /comandes/{comandaId}/linies (capa 30) nunca lo
            consulta — siempre usa la tarifa del cliente.
        dataLliurament: { type: string, format: date-time, nullable: true }
        transportistaId: { type: integer, nullable: true }
        obsLliurament: { type: string, nullable: true }
        linies:
          type: array
          items: { $ref: '#/components/schemas/LiniaEntrada' }
    LiniaEntrada:
      type: object
      description: >-
        Una línea nueva — usada tanto embebida en ComandaEntrada (alta
        completa) como en el body de POST /comandes/{comandaId}/linies
        (capa 30, agregar línea a un pedido ya existente). Mismo shape en
        los dos casos.
      required: [producteId, unitatsDemanades]
      properties:
        producteId: { type: integer }
        unitatsDemanades:
          type: number
          exclusiveMinimum: 0
          multipleOf: 0.01
          description: >-
            Capa 38. JS number normal en la entrada, hasta 2 decimales
            (NUMERIC(10,2), entregas/pedidos parciales de pieza) — más de 2
            decimales rechaza con 400 VALIDACIO.
        kgDemanats:
          type: string
          nullable: true
          description: Sólo si el artículo es «a medida» (sin pesKg)
        dataProduccio:
          type: string
          format: date-time
          nullable: true
          description: >-
            Capa 34. Opcional. Se valida contra las 6 reglas de coherencia
            temporal (ComandaEntradaParcial.dataProduccio) — en la alta
            (ComandaEntrada) sólo contra dataLliurament, la única fecha de
            cabecera que existe en ese body; en
            POST /comandes/{comandaId}/linies contra las 3 fechas de
            cabecera ya guardadas del pedido.
    ComandaEntradaParcial:
      type: object
      properties:
        tarifaId: { type: integer, nullable: true }
        transportistaId: { type: integer, nullable: true }
        poblacioDesti: { type: string, nullable: true }
        adrecaLliurament: { type: string, nullable: true }
        dataProduccio:
          type: string
          format: date-time
          nullable: true
          description: >-
            Capa 34. Si cambia cualquiera de dataProduccio/dataExpedicio/
            dataLliurament, se valida el estado resultante (valor nuevo si
            vino, si no el ya guardado) contra las 6 reglas de coherencia
            temporal (ver docs/contrato-api.md § 4.5) — no sólo contra las
            otras 2 fechas de cabecera (reglas 1/2/3), sino también contra
            TODAS las líneas activas del pedido (reglas 4/5/6), aunque
            ninguna se esté tocando en este mismo request.
        dataExpedicio: { type: string, format: date-time, nullable: true }
        dataLliurament: { type: string, format: date-time, nullable: true }
        bultos: { type: integer, nullable: true }
        obsProduccio: { type: string, nullable: true }
        obsLliurament: { type: string, nullable: true }
        estat:
          allOf: [{ $ref: '#/components/schemas/EstatComanda' }]
          description: >-
            Capa 31. Transición libre entre los 5 estados, sin máquina de
            estados. "cancellada" saca el pedido de todos los paneles. Si el valor es "amb_incidencia", `detall` es obligatorio
            en el mismo body y se registra como una incidencia nueva
            (tipus "manual") — a diferencia de las incidencias automáticas
            (sense_preu, etc.), ésta la dispara un usuario a mano.
        detall:
          type: string
          nullable: true
          description: >-
            Sólo se usa (y es obligatorio) cuando estat = "amb_incidencia".
            Se ignora en cualquier otro caso.
...
    Usuari:
      type: object
      description: >-
        Capa 17. Se crea por auto-provisioning (GET /jo y cualquier otro
        endpoint de negocio) o por alta manual de un Administrador
        (POST /usuaris, capa 19).
      properties:
        id: { type: integer }
        firebaseUid:
          type: string
          description: uid de Firebase Auth — vínculo con el token, ver ADR-021.
        nom: { type: string }
        email: { type: string }
        rol:
          type: object
          description: >-
            Incluye modulsPermesos acá (no sólo en el schema Rol) porque
            GET /jo es lo primero que llama el frontend y necesita saberlo
            sin una segunda llamada a GET /rols/{id}.
          properties:
            id: { type: integer }
            nom: { type: string }
            modulsPermesos:
              type: array
              items: { type: string }
        actiu: { type: boolean }
    UsuariCreacio:
      type: object
      description: >-
        Cuerpo de POST /usuaris (capa 19). A diferencia del auto-
        provisioning, no lleva firebaseUid: lo genera el backend al crear
        el usuario en Firebase, no lo elige quien da de alta.
      required: [nom, email, rolId]
      properties:
        nom: { type: string }
        email: { type: string }
        rolId: { type: integer }
    UsuariCreatResposta:
      type: object
      description: >-
        Respuesta de POST /usuaris (capa 19). No hay envío de email — el
        Administrador comparte linkEstabliment a mano por el canal que use.
      properties:
        usuari: { $ref: '#/components/schemas/Usuari' }
        linkEstabliment:
          type: string
          description: >-
            Link de Firebase (generatePasswordResetLink) de un solo uso
            para que la persona establezca su propia contraseña. Nunca debe
            loggearse ni guardarse más allá de esta respuesta.
    UsuariEdicio:
      type: object
      description: >-
        Cuerpo de PATCH /usuaris/{id} — sólo los campos editables.
        firebaseUid y email son inmutables, no están acá a propósito.
      properties:
        nom: { type: string }
        rolId: { type: integer }
        actiu: { type: boolean }